Cast action to float32

+++ dlr_sara_grid_clamp_dataset/dlr_sara_grid_clamp_dataset_dataset_builder.py
@@ -109,7 +109,7 @@ class DlrSaraGridClampDataset(tfds.core.GeneratorBasedBuilder):
                             # 'wrist_image': step['wrist_image'],
                             "state": step["state"],
                         },
-                        "action": step["action"],
+                        "action": step["action"].astype(np.float32),
                         "discount": 1.0,
                         "reward": step["reward"],
                         "is_first": i == 0,
